fix: drop redundant PortNumber conversions flagged by unconvert
Gateway API 1.6 made PortNumber an int32 alias, so the explicit conversions
around listener.Port are now no-ops and fail the unconvert linter. Same class as
the `port = *backend.Port` change master already made in TranslateHTTPRoute.

+++ internal/adc/translator/grpcroute.go
@@ -312,7 +312,7 @@ func (t *Translator) TranslateGRPCRoute(tctx *provider.TranslateContext, grpcRou
 		// Collect unique listener ports for port-based routing.
 		listenerPorts := make(map[int32]struct{})
 		for _, listener := range tctx.Listeners {
-			listenerPorts[int32(listener.Port)] = struct{}{}
+			listenerPorts[listener.Port] = struct{}{}
 		}
 
 		if t.shouldInjectServerPortVars(tctx.RouteParentRefs, listenerPorts) {
